src/main/java/servlet/LogOutServlet.java
package servlet;

import accounts.AccountService;
import accounts.FactoryAccountService;
import org.apache.logging.log4j.LogManager;
import org.apache.logging.log4j.Logger;
import utils.PageMessageUtil;

import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

@WebServlet(name = "LogOut", urlPatterns = "/logout")
public class LogOutServlet extends HttpServlet {
    private AccountService accountService;
    static final Logger LOGGER = LogManager.getLogger(LogOutServlet.class.getName());

    @Override
    public void init() throws ServletException {
        super.init();
        this.accountService = FactoryAccountService.getAccountService();
    }

    @Override
    public void destroy() {
        super.destroy();
        this.accountService = null;
    }

    @Override
    public void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        LOGGER.debug("doPost from " + this.getClass().getSimpleName());

        PageMessageUtil.clearPageMessageForDoPost(request);

        String sessionId = request.getSession().getId();
        accountService.deleteSession(sessionId);

        response.setContentType("text/html;charset=utf-8");
        response.sendRedirect(request.getContextPath() + SignInServlet.URL);
    }
}
